004-005-model-factory-and-seeder

+++ database/migrations/2023_01_20_204221_add_fields_to_listings_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             //
             $table->unsignedTinyInteger('beds');
             $table->unsignedTinyInteger('baths');
-            $table->unsignedTinyInteger('area');
+            $table->unsignedSmallInteger('area');
 
             $table->tinyText('city');
             $table->tinyText('code');
